moving frame move constructor from cpp to h, for gl use
